python/requirements.txt
Flask==2.2.2
python-mecab-ko